move put integration test to correct file
